chore: update version to 2.0.1 and refactor configuration model references
- Updated version to 2.0.1 in pyproject.toml.
- Refactored code to replace references from Configuration to ConfigurationModel across multiple files.
- Added new configuration_model.py file to define the ConfigurationModel class.

This update improves code organization and aligns with the new model structure.
