+++ qase-api-client/src/qase/api_client_v1/models/configuration_group.py
@@ -20,7 +20,7 @@ import json
 
 from pydantic import BaseModel, ConfigDict, StrictInt, StrictStr
 from typing import Any, ClassVar, Dict, List, Optional
-from qase.api_client_v1.models.configuration import ConfigurationModel
+from qase.api_client_v1.models.configuration_model import ConfigurationModel
 from typing import Optional, Set
 from typing_extensions import Self
