+++ qase-api-client/src/qase/api_client_v1/models/configuration_model.py
